pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "3.13.11"
fastapi = "^0.128.0"
uvicorn = {version = "^0.40.0", extras = ["standard"]}
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.45"}
asyncpg = "^0.31.0"
alembic = "^1.18.1"
pybase62 = "^1.0.0"
httpx = "^0.28.1"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
